requirements.txt
pydantic[email]==1.10.7
pydantic==1.10.7
swagger-ui-py==22.7.13
Flask==2.2.2
requests~=2.28.1
msal~=1.20.0
responses~=0.23.1
pytest~=7.2.2
prance~=0.22.11.4.0
openapi~=1.1.0
openapi-spec-validator~=0.5.6
jsf~=0.7.1
pillow~=10.0.0
scandir~=1.10.0
backports.shutil-get-terminal-size~=1.0.0
browsepy~=0.5.6
art
flake8~=4.0.1
black~=22.3.0
types-requests==2.31.0.1
bandit~=1.7.2